requirements.txt
flask==3.0.3
gunicorn==22.0.0
pillow==10.4.0
numpy>=1.26
ultralytics>=8.3.0
opencv-python-headless>=4.9.0.80
torch>=2.2.0
